coldopen-coach/requirements.txt
mcp>=1.0.0
apify-client>=1.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
typing-extensions>=4.8.0